test(agent): share the scheduler-auth and message-signing helpers
scheduler_auth lived in test_vprogram.py and hardcoded the legacy path into
the signed payload, which the signature covers, so reusing it against a v2
route answered 401. It moves to conftest with the path as a parameter and
the legacy path as its default. sign_message and the instance content it
signs move with it, since the v2 tests need a genuinely signed message to
break; test_allocation_verify keeps its fixtures and reads them from there.
